server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.123.7
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
frozenlist==1.8.0
genai-prices==0.0.47
google-auth==2.43.0
google-genai==1.55.0
greenlet==3.2.4
griffe==1.15.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
logfire-api==4.16.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
ngrok==1.4.0
opentelemetry-api==1.39.1
orjson==3.11.4
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.10
pwdlib==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-ai-slim==1.31.0
pydantic-extra-types==2.10.6
pydantic-graph==1.31.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.7
pytest==9.0.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.47.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
tenacity==9.1.2
twilio==9.8.8
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
